Edited to support data in home directory
Made a few changes to the code so that it initially checks for the `.todo-list.json` file in your home directory `~/.config/todo/`, then (*as the original*) in the project's root, then in parent directories.  I did this to support a multi-user environment, as well as to align with my existing Linux endpoint configuration practices.
Then added a little text in the readme "Installation" section to align with this code change.